ex 10

+++ src/app.ts
@@ -11,6 +11,7 @@ import { RentNotFoundError } from "./errors/rent-not-found-error";
 import { RentRepo } from "./ports/rent-repo";
 import { UserRepo } from "./ports/user-repo";
 import { BikeRepo } from "./ports/bike-repo";
+import { UserHasOpenRent } from "./errors/user-has-open-rent";
 
 export class App {
     crypt: Crypt = new Crypt()
@@ -19,7 +20,7 @@ export class App {
         readonly userRepo: UserRepo,
         readonly bikeRepo: BikeRepo,
         readonly rentRepo: RentRepo
-    ) {}
+    ) { }
 
     async findUser(email: string): Promise<User> {
         const user = await this.userRepo.find(email)
@@ -29,7 +30,7 @@ export class App {
 
     async registerUser(user: User): Promise<string> {
         if (await this.userRepo.find(user.email)) {
-          throw new DuplicateUserError()
+            throw new DuplicateUserError()
         }
         const encryptedPassword = await this.crypt.encrypt(user.password)
         user.password = encryptedPassword
@@ -46,10 +47,11 @@ export class App {
     }
 
     async removeUser(email: string): Promise<void> {
+        if((await this.rentRepo.findOpenRentsFor(email)).length > 0) throw new UserHasOpenRent()
         await this.findUser(email)
         await this.userRepo.remove(email)
     }
-    
+
     async rentBike(bikeId: string, userEmail: string): Promise<string> {
         const bike = await this.findBike(bikeId)
         if (!bike.available) {
@@ -97,7 +99,7 @@ export class App {
 }
 
 function diffHours(dt2: Date, dt1: Date) {
-  var diff = (dt2.getTime() - dt1.getTime()) / 1000;
-  diff /= (60 * 60);
-  return Math.abs(diff);
+    var diff = (dt2.getTime() - dt1.getTime()) / 1000;
+    diff /= (60 * 60);
+    return Math.abs(diff);
 }
